+++ example/lib/use_cases/installed_apps_page.dart
@@ -56,6 +56,7 @@ class _InstalledAppsPageState extends State<InstalledAppsPage> {
     _InstalledItem('Maps', 'Apple Maps', AppleMaps(), 'assets/apple_maps.png'),
     _InstalledItem('Maps', 'Sygic', Sygic(), 'assets/sygic.png'),
     _InstalledItem('Maps', 'Moovit', Moovit(), 'assets/moovit.png'),
+    _InstalledItem('Maps', 'Air Navigation Pro', AirNavigationPro(), 'assets/air_navigation_pro.png'),
     _InstalledItem('Maps', 'Neshan', Neshan(), 'assets/neshan.png'),
     _InstalledItem('Maps', 'Yandex Maps', YandexMaps(), 'assets/yandex_maps.png'),
     _InstalledItem('Maps', 'Yandex Navigator', YandexNavigator(), 'assets/yandex_navigator.png'),