+++ lib/src/apps/downloadable_apps/downloadable_apps.dart
@@ -1,3 +1,4 @@
+export 'air_navigation_pro.dart';
 export 'amap.dart';
 export 'apple_maps.dart';
 export 'baidu_maps.dart';